fix input fn

+++ pretrain/pretrain_data.py
@@ -47,7 +47,7 @@ def get_input_fn(config: configure_pretraining.PretrainingConfig, is_training,
         "segment_ids": tf.io.FixedLenFeature([config.max_seq_length], tf.int64),
         "masked_lm_positions": tf.io.FixedLenFeature([config.max_seq_length], tf.int64),
         "masked_lm_ids": tf.io.FixedLenFeature([config.max_seq_length], tf.int64),
-        "masked_lm_weights": tf.io.FixedLenFeature([config.max_seq_length], tf.int64),
+        "masked_lm_weights": tf.io.FixedLenFeature([config.max_seq_length], tf.float32),
     }
 
     d = tf.data.Dataset.from_tensor_slices(tf.constant(input_files))
